+++ setup.cfg
@@ -67,7 +67,7 @@ glotaran.plugins.megacomplexes =
     spectral = glotaran.builtin.megacomplexes.spectral
 glotaran.plugins.project_io =
     yml = glotaran.builtin.io.yml.yml
-    csv = glotaran.builtin.io.csv.csv
+    csv = glotaran.builtin.io.pandas.csv
     folder = glotaran.builtin.io.folder.folder_plugin
 
 [aliases]
